pymatch stopped working with pandas 0.24
Hi Folks!
First of all, thanx for this library, it has been of great help to me! :-D

Recently, after a pandas update (0.24 I think) it stopped working saying that it couldn't coerce to Series due to dimentions missmatch. I managed to fix it by adding some square brackets to Matcher's function _scores_to_accuracy.

Also, it had some nasty warnings about pandas concat and append functions that dissappeared with a sort=True parameter.

Next, I noticed that the __init__ function dropped all rows that had a null on it, without checking if that column was actually part of the PSM process, so I restricted that dropna only to xvar columns.

And at last and least, I added an informative message "Static column dropped:" to drop_static_cols just as a courtesy to the user.

Attached to this email, you'll find the modified files for your review and evaluation for the next release.

Atentamente: / Best Regards: / Atenciosamente: / Cordialement:

Alejandro Gutiérrez González